requirements.txt
streamlit==1.31.0
tensorflow==2.17.0
Pillow==10.4.0